+++ docs/release_notes.rst
@@ -10,7 +10,7 @@ releases are available on `Anaconda.org
 0.0.6 - 2020-
 -------------
 
--
+- Add ``run_always`` to the task dictionary to always execute a task (:gh:`12`).
 
 
 0.0.5 - 2020-04-26