+++ docs/release_notes.rst
@@ -10,6 +10,7 @@ releases are available on `Anaconda.org
 0.0.6 - 2020-
 -------------
 
+- Add ``run_always`` to the task dictionary to always execute a task (:gh:`12`).
 - Add project root directory to the ``PYTHONPATH`` so that every task can import with
   ``from src.... import ...`` (:gh:`13`).
 - Allow to have directories as task dependencies and targets (:gh:`15`).